Исправил замечания Убрал точку входа вya_file_manager.py упаковал переменную со списком для хендлера клавиатуры в сам хендлер запрятал set_correct_path в классе
